Refactor HistoricalWinning
* Refactor HistoricalWinning to FantasyLeagues.HistoricalWinning
* See #797
