+++ lib/ex338_web/controllers/page_controller.ex
@@ -1,16 +1,13 @@
 defmodule Ex338Web.PageController do
   use Ex338Web, :controller
 
-  alias Ex338.{
-    FantasyLeagues,
-    HistoricalWinning
-  }
+  alias Ex338.{FantasyLeagues}
 
   def index(conn, _params) do
     leagues = FantasyLeagues.get_leagues_by_status("primary")
     season_records = FantasyLeagues.get_current_season_records()
     all_time_records = FantasyLeagues.get_current_all_time_records()
-    winnings = HistoricalWinning.Store.get_all_winnings()
+    winnings = FantasyLeagues.get_all_winnings()
 
     render(conn, "index.html",
       all_time_records: all_time_records,
